refactor/#45 : DTO 이동
