+++ api/src/main/java/io/eagle/domain/transaction/repository/TransactionRepositoryCustom.java
@@ -1,7 +1,7 @@
 package io.eagle.domain.transaction.repository;
 
-import io.eagle.domain.transaction.dto.TransactionRequestDto;
-import io.eagle.domain.transaction.dto.TransactionResponseDto;
+import io.eagle.domain.transaction.dto.request.TransactionRequestDto;
+import io.eagle.domain.transaction.dto.response.TransactionResponseDto;
 import io.eagle.entity.Order;
 import io.eagle.entity.Transaction;
 import org.springframework.data.domain.Pageable;
