+++ api/src/main/java/io/eagle/domain/transaction/repository/TransactionRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.querydsl.core.types.dsl.DateTemplate;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.JPQLQueryFactory;
 import io.eagle.domain.transaction.dto.QTransactionResponseDto;
-import io.eagle.domain.transaction.dto.TransactionRequestDto;
-import io.eagle.domain.transaction.dto.TransactionResponseDto;
+import io.eagle.domain.transaction.dto.request.TransactionRequestDto;
+import io.eagle.domain.transaction.dto.response.TransactionResponseDto;
 import io.eagle.entity.Order;
 import io.eagle.entity.Transaction;
 import lombok.RequiredArgsConstructor;
