+++ api/src/test/java/io/eagle/domain/transaction/repository/TransactionRepositoryTest.java
@@ -3,8 +3,8 @@ package io.eagle.domain.transaction.repository;
 import io.eagle.common.TestUtil;
 import io.eagle.config.TestConfig;
 import io.eagle.domain.order.repository.OrderRepository;
-import io.eagle.domain.transaction.dto.TransactionRequestDto;
-import io.eagle.domain.transaction.dto.TransactionResponseDto;
+import io.eagle.domain.transaction.dto.request.TransactionRequestDto;
+import io.eagle.domain.transaction.dto.response.TransactionResponseDto;
 import io.eagle.domain.vacation.repository.VacationRepository;
 import io.eagle.entity.Order;
 import io.eagle.entity.Transaction;
